Refactor RecallValidator
