--- modules/recheck-core/js/src/main/scala/codes/quine/labo/recheck/recall/RecallValidator.scala
@@ -1,18 +1,0 @@
-package codes.quine.labo.recheck.recall
-
-import scala.annotation.nowarn
-import scala.concurrent.duration.Duration
-
-import codes.quine.labo.recheck.common.Context
-import codes.quine.labo.recheck.common.UnexpectedException
-import codes.quine.labo.recheck.diagnostics.AttackPattern
-
-object RecallValidator {
-  @nowarn
-  def validate(source: String, flags: String, pattern: AttackPattern, timeout: Duration)(implicit
-      ctx: Context
-  ): RecallResult = timeout match {
-    case d if d < Duration.Zero => RecallResult.Timeout
-    case _                      => throw new UnexpectedException("recall validation is not supported.")
-  }
-}
